# specs/001-goal-build-a/contracts/api-contracts.yaml
openapi: 3.1.0
info:
  title: Brand Analysis & Scoring Platform API
  version: 1.0.0
  description: |
    REST API for brand analysis platform. Enables brand creation, prompt management,
    competitor tracking, and AI-powered scoring across 5 dimensions.
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://brand-analysis.vercel.app
    description: Production
...
paths:
...
  /api/brands:
    post:
      tags: [Brands]
      summary: Create a new brand
      description: |
        Creates a new brand profile (FR-001). Validates domain format (FR-002).
        After creation, user should be redirected to prompts page (FR-003).
      operationId: createBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBrandRequest'
            examples:
              valid:
                value:
                  brand_name: "EcoClean"
                  domain: "ecoclean.com"
                  region: "Global"
      responses:
        '201':
          description: Brand created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
              examples:
                success:
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    brand_name: "EcoClean"
                    domain: "ecoclean.com"
                    region: "Global"
                    created_at: "2025-10-15T10:30:00Z"
                    updated_at: "2025-10-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Domain already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  value:
                    error: "A brand with this domain already exists"
                    code: "DUPLICATE_DOMAIN"
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  schemas:
...
    CreateBrandRequest:
      type: object
      required:
        - brand_name
        - domain
        - region
      properties:
        brand_name:
          type: string
          minLength: 1
          maxLength: 255
          description: Brand or company name
          example: "EcoClean"
        domain:
          type: string
          pattern: '^[a-zA-Z0-9][a-zA-Z0-9-]{0,61}[a-zA-Z0-9]?\.[a-zA-Z]{2,}$'
          description: Website domain (e.g., example.com)
          example: "ecoclean.com"
        region:
          type: string
          minLength: 1
          maxLength: 100
          description: Geographic region
          example: "Global"
    BrandResponse:
      type: object
      required:
        - id
        - brand_name
        - domain
        - region
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        brand_name:
          type: string
        domain:
          type: string
        region:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        details:
          type: object
          description: Additional error context
          additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                error: "Validation failed"
                code: "VALIDATION_ERROR"
                details:
                  field: "domain"
                  message: "Invalid domain format"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              value:
                error: "An unexpected error occurred"
                code: "INTERNAL_ERROR"